feat : (CMake file to build from VS Code)
