--- Derivatives/c++/Options/Options.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30804.86
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Options", "Options\Options.vcxproj", "{20C9170F-90A3-4FAA-BCC2-3F141C905B46}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{20C9170F-90A3-4FAA-BCC2-3F141C905B46}.Debug|x64.ActiveCfg = Debug|x64
-		{20C9170F-90A3-4FAA-BCC2-3F141C905B46}.Debug|x64.Build.0 = Debug|x64
-		{20C9170F-90A3-4FAA-BCC2-3F141C905B46}.Debug|x86.ActiveCfg = Debug|Win32
-		{20C9170F-90A3-4FAA-BCC2-3F141C905B46}.Debug|x86.Build.0 = Debug|Win32
-		{20C9170F-90A3-4FAA-BCC2-3F141C905B46}.Release|x64.ActiveCfg = Release|x64
-		{20C9170F-90A3-4FAA-BCC2-3F141C905B46}.Release|x64.Build.0 = Release|x64
-		{20C9170F-90A3-4FAA-BCC2-3F141C905B46}.Release|x86.ActiveCfg = Release|Win32
-		{20C9170F-90A3-4FAA-BCC2-3F141C905B46}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {94C3B7CE-9F3A-4EE4-8E42-6AFF7D7A0F1A}
-	EndGlobalSection
-EndGlobal
